Rename to state-ui-lib
