+++ todomvc/components/AddArea.js
@@ -3,7 +3,7 @@
 //     <button>Add Todo<button>
 // <form/>
 
-import { dom as d } from '../lib/state-manager';
+import { dom as d } from '../lib/state-ui-lib';
 
 function AddArea({ onAddButtonClick }) {
   function addButtonClickHandler(event) {
