+++ todomvc/lib/state-ui-lib/pubsub.js
